this is all loking difficult
I need to be able to check if the latest release matches the name of a
folder in my bin dir. If it doesnt, then delete the entire folder in
case there are old bins. This means the binary does need to be
downloaded and extracted into the directory we checked for.